requirements.txt
Django==4.1.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
python-decouple==3.7
pytz~=2023.3.post1
PyJWT~=2.8.0